+++ src/main/java/umc/cockple/demo/domain/chat/converter/ChatConverter.java
@@ -9,6 +9,7 @@ import umc.cockple.demo.domain.chat.enums.WebSocketMessageType;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.chat.dto.*;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -113,6 +114,20 @@ public class ChatConverter {
                 .senderId(sender.getId())
                 .senderName(sender.getMemberName())
                 .senderProfileImageUrl(senderProfileImageUrl)
+                .createdAt(savedMessage.getCreatedAt())
+                .build();
+    }
+
+    public WebSocketMessageDTO.Response toSystemMessageResponse(
+            Long chatRoomId, String content, ChatMessage savedMessage){
+        return WebSocketMessageDTO.Response.builder()
+                .type(WebSocketMessageType.SEND)
+                .chatRoomId(chatRoomId)
+                .messageId(savedMessage.getId())
+                .content(content)
+                .senderId(null)
+                .senderName("시스템")
+                .senderProfileImageUrl(null)
                 .createdAt(savedMessage.getCreatedAt())
                 .build();
     }
