+++ src/main/java/umc/cockple/demo/domain/chat/events/ChatEventListener.java
@@ -16,16 +16,16 @@ import umc.cockple.demo.domain.party.events.PartyMemberJoinedEvent;
 @Slf4j
 public class ChatEventListener {
 
-    private final ChatRoomService chatRoomService;
+    private final ChatWebSocketService chatWebSocketService;
 
     @EventListener
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT) //트랜잭션이 커밋된 후에 실행
     @Async
     public void handlePartyMemberChanged(PartyMemberJoinedEvent event) {
         switch (event.action()) {
-            case JOINED -> chatRoomService.sendSystemMessage(event.partyId(),
+            case JOINED -> chatWebSocketService.sendSystemMessage(event.partyId(),
                     event.memberName() + "님이 모임에 참여하셨습니다.");
-            case LEFT -> chatRoomService.sendSystemMessage(event.partyId(),
+            case LEFT -> chatWebSocketService.sendSystemMessage(event.partyId(),
                     event.memberName() + "님이 모임을 떠나셨습니다.");
         }
     }
